requirements.txt
streamlit
pandas
requests
google-generativeai
scikit-surprise